pyproject.toml
[project]
name = "mockins-weekly-summary"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "openpyxl>=3.1",
  "python-dotenv>=1.0",
  "pydantic>=2.7",
  "requests>=2.32",
  "google-api-python-client>=2.120",
  "google-auth>=2.29",

]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-requests>=2.32",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
